Integrated optimizing compiler into Decapodes
Also added AdvDiff and Brusselator simulations to test new compilation
