+++ examples/Testing Range/Brusselator.jl
@@ -29,8 +29,8 @@ function generate(sd, my_symbol; hodge=GeometricHodge())
 end
 
 begin
-    # primal_earth = loadmesh(Icosphere(5))
-    primal_earth = EmbeddedDeltaSet2D("Icosphere8.obj")
+    primal_earth = loadmesh(Icosphere(3))
+    # primal_earth = EmbeddedDeltaSet2D("Icosphere8.obj")
     nploc = argmax(x -> x[3], primal_earth[:point])
     primal_earth[:edge_orientation] .= false
     orient!(primal_earth)
@@ -57,17 +57,7 @@ begin
         ∂ₜ(V) == V̇
     end))
 
-    bespoke_op1_inf_rules = [
-        (src_type = :Form0, tgt_type = :infer, replacement_type = :Form0, op = :Δ)]
-  
-    bespoke_op2_inf_rules = [
-        (proj1_type = :Form0, proj2_type = :Form0, res_type = :infer, replacement_type = :Form0, op = :.*),
-        (proj1_type = :Form0, proj2_type = :Parameter, res_type = :infer, replacement_type = :Form0, op = :*),
-        (proj1_type = :Parameter, proj2_type = :Form0, res_type = :infer, replacement_type = :Form0, op = :*)]
-  
-    infer_types!(Brusselator,
-        vcat(bespoke_op1_inf_rules, op1_inf_rules_2D),
-        vcat(bespoke_op2_inf_rules, op2_inf_rules_2D))
+    infer_types!(Brusselator)
 
     resolve_overloads!(Brusselator)
 end
@@ -118,11 +108,11 @@ end
 end =#
 
 sim = eval(gensim(Brusselator))
-sim = eval_gensim(Brusselator, CUDA=true, no_infers=true,
+#= sim = eval_gensim(Brusselator, CUDA=true, no_infers=true,
     verbose_errors=true, precision=Float64,
-    primals=[:J, :B], duals=[:E])
+    primals=[:J, :B], duals=[:E]) =#
 fₘ = sim(earth, generate)
-fₘ = sim(earth)
+# fₘ = sim(earth)
 # fₘ = simulate(earth, generate)
 
 begin
@@ -150,7 +140,7 @@ begin
     u₀ = construct(PhysicsState, [VectorForm(U), VectorForm(V)],Float64[], [:U, :V])
     tₑ = 11.5
     prob = ODEProblem(fₘ,u₀,(0, tₑ), constants_and_parameters)
-    soln = solve(prob, Tsit5(), save_everystep = false)
+    soln = solve(prob, Tsit5())
 end
 
 fig, ax, ob = GLMakie.mesh(primal_earth, color = findnode(soln(0), :U))
