+++ test/Decapodes2/multiscalearrays.jl
@@ -107,9 +107,9 @@ end
 
 advdiff = parse_decapode(AdvDiff)
 advdiffdp = SummationDecapode(advdiff)
-Decapodes.compile(advdiffdp, [:C, :V])
-Decapodes.compile(expand_operators(advdiffdp), [:C, :V])
-gensim(expand_operators(advdiffdp), [:C, :V])
+# Decapodes.compile(advdiffdp, [:C, :V])
+# Decapodes.compile(expand_operators(advdiffdp), [:C, :V])
+# gensim(expand_operators(advdiffdp), [:C, :V])
 sim = eval(gensim(expand_operators(advdiffdp), [:C, :V]))
 fₘ = sim(periodic_mesh, generate)
 velocity(p) = [-0.5, -0.5, 0.0]
